requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
neo4j==5.12.0
langchain==0.0.291
langchain_openai==0.0.2
openai==1.2.4
sentence-transformers==2.2.2
transformers==4.34.0
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.31.0
numpy==1.26.0
pandas==2.1.1
torch==2.1.0
python-multipart==0.0.6
google-generativeai>=0.3.0